refactor(vxtwitter): use `#![feature(let_chains)]` to simplify handling

+++ src/vxtwitter.rs
@@ -17,28 +17,24 @@ pub async fn handler(
     bot: AutoSend<Bot>,
     message: Message,
 ) -> Result<(), Box<dyn Error + Sync + Send + 'static>> {
-    if let Some(text) = message.text() {
-        if let Some(user) = message.from() {
-            if TWITTER_REGEX.is_match(text) {
-                let text = text.replace("https://twitter.com", "https://vxtwitter.com");
-                let text = format!(
-                    "<a href=\"{}\">{}</a>: {}",
-                    user.id.url(),
-                    user.full_name(),
-                    text
-                );
-                bot.delete_message(message.chat.id, message.id).await?;
-                if let Some(reply) = message.reply_to_message() {
-                    bot.send_message(message.chat.id, text)
-                        .reply_to_message_id(reply.id)
-                        .parse_mode(ParseMode::Html)
-                        .await?;
-                } else {
-                    bot.send_message(message.chat.id, text)
-                        .parse_mode(ParseMode::Html)
-                        .await?;
-                }
-            }
+    if let Some(text) = message.text() && let Some(user) = message.from() && TWITTER_REGEX.is_match(text) {
+        let text = text.replace("https://twitter.com", "https://vxtwitter.com");
+        let text = format!(
+            "<a href=\"{}\">{}</a>: {}",
+            user.id.url(),
+            user.full_name(),
+            text
+        );
+        bot.delete_message(message.chat.id, message.id).await?;
+        if let Some(reply) = message.reply_to_message() {
+            bot.send_message(message.chat.id, text)
+                .reply_to_message_id(reply.id)
+                .parse_mode(ParseMode::Html)
+                .await?;
+        } else {
+            bot.send_message(message.chat.id, text)
+                .parse_mode(ParseMode::Html)
+                .await?;
         }
     }
     Ok(())
